allom.predict

Function for making tree-level Monte Carlo predictions from allometric equations estimated from the PEcAn allometry module

Usage

allom.predict(
  object,
  dbh,
  pft = NULL,
  component = NULL,
  n = NULL,
  use = "Bg",
  interval = "prediction",
  single.tree = FALSE
)

Arguments

object

Allometry model object. Option includes

'list of mcmc'

- mcmc outputs in a list by PFT then component

'vector of file paths'

- path(s) to AllomAve RData files

'directory where files are located

-

dbh

Diameter at Breast Height (cm)

pft

Plant Functional Type. Needs to match the name used in AllomAve. Can be NULL if only one PFT/species exists, otherwise needs to the same length as dbh

component

Which component to predict. Can be NULL if only one component was analysed in AllomAve.

n

Number of Monte Carlo samples. Defaults to the same number as in the MCMC object

use

c('Bg','mu','best')

interval

c('none','confidence','prediction') default is prediction

single.tree

logical: Is this a DBH time series from one individual tree? If TRUE, will use a fixed error for all draws.

Value

matrix of Monte Carlo predictions that has n rows and one column per DBH

Author

Michael Dietze, Christy Rollinson

Examples

if (FALSE) { # \dontrun{
  object = '~/Dropbox//HF C Synthesis/Allometry Papers & Analysis/'
  dbh = seq(10,50,by=5)
  mass = allom.predict(object,dbh,n=100)

} # }